Moved the geospatial data from DB2 to H2. Re-factored the file system handler. Fixed the index context serialization.
git-svn-id: http://ikube.googlecode.com/svn/trunk@786 89a0fb20-889e-aefa-90f6-7c296b20d5cf

+++ code/core/src/main/java/ikube/model/Server.java
@@ -36,7 +36,7 @@ public class Server extends Persistable implements Comparable<Server> {
 	private long age;
 
 	private String searchWebServiceUrl;
-	
+
 	public Server() {
 		this.actions = new ArrayList<Action>();
 	}
@@ -52,8 +52,6 @@ public class Server extends Persistable implements Comparable<Server> {
 	public String getAddress() {
 		return address;
 	}
-	
-	
 
 	public List<Action> getActions() {
 		return actions;
@@ -124,7 +122,7 @@ public class Server extends Persistable implements Comparable<Server> {
 		if (this.getId() == 0) {
 			return super.hashCode();
 		}
-		return (int) (long) this.getId();
+		return (int) this.getId();
 	}
 
 	@Override
